refactor: DefaultFareCalculator 내 @Service 삭제 및 Calculator 관련 객체 domain 패키지로 이동
