+++ src/main/java/wooteco/subway/domain/FareCalculator.java
@@ -1,4 +1,4 @@
-package wooteco.subway.service;
+package wooteco.subway.domain;
 
 public interface FareCalculator {
     int calculateFare(int distance, int extraFare);
